requirements.txt
imap-tools>=1.0.0
python-docx>=0.8.11
ffmpeg-normalize>=1.24.0
